Added maze example

+++ src/environment/finite.rs
@@ -4,26 +4,26 @@ use environment::{Space, FiniteSpace};
 
 #[derive(Debug, Clone, Copy)]
 pub struct Finite {
-	size: u64,
+	size: usize,
 }
 
 impl Space for Finite {
-	type Element = u64;
+	type Element = usize;
 
-	fn sample(&self) -> u64 {
+	fn sample(&self) -> usize {
 		let mut rng = thread_rng();
 		rng.gen_range(0, self.size)
 	}
 }
 
 impl FiniteSpace for Finite {
-	fn enumerate(&self) -> Vec<u64> {
+	fn enumerate(&self) -> Vec<usize> {
 		(0..self.size).collect()
 	}
 }
 
 impl Finite {
-	pub fn new(size: u64) -> Finite {
+	pub fn new(size: usize) -> Finite {
 		Finite {
 			size: size
 		}
